ESLint fix

+++ src/provider.js
@@ -1,38 +1,78 @@
+import ModelPoint from './model-point.js';
+
+const objectToArray = (object) => {
+  return Object.keys(object).map((id) => object[id]);
+};
+
 const Provider = class {
-  constructor({api, store}) {
+  constructor({api, store, generateId}) {
     this._api = api;
     this._store = store;
+    this._generateId = generateId;
+    this._needSync = false;
+  }
+
+  _isOnline() {
+    return window.navigator.onLine;
   }
 
   updatePoint({id, data}) {
-    return this._api.updatePoint({id, data})
-      .then((point) => {
-        this._store.setItem({key: point.id, item: point.toRAW()});
-        return point;
-      });
+    if (this._isOnline()) {
+      return this._api.updatePoint({id, data})
+        .then((point) => {
+          this._store.setItem({key: point.id, item: point.toRAW()});
+          return point;
+        });
+    } else {
+      const point = data;
+      this._needSync = true;
+      this._store.setItem({key: point.id, item: point});
+      return Promise.resolve(ModelPoint.parsePoint(point));
+    }
   }
 
   createPoint({point}) {
-    return this._api.createPoint({point})
-      .then((point) => {
-        this._store.setItem({key: point.id, item: point.toRAW()});
-        return point;
-      });
+    if (this._isOnline()) {
+      return this._api.createPoint({point})
+        .then((newPoint) => {
+          this._store.setItem({key: newPoint.id, item: newPoint.toRAW()});
+          return point;
+        });
+    } else {
+      point.id = this._generateId();
+      this._needSync = true;
+
+      this._store.setItem({key: point.id, item: point});
+      return Promise.resolve(ModelPoint.parsePoint(point));
+    }
   }
 
   deletePoint({id}) {
-  return this._api.deletePoint({id})
-    .then(() => {
+    if (this._isOnline()) {
+      return this._api.deletePoint({id})
+        .then(() => {
+          this._store.removeItem({key: id});
+        });
+    } else {
+      this._needSync = true;
       this._store.removeItem({key: id});
-    });
-}
+      return Promise.resolve(true);
+    }
+  }
 
   getPoints() {
-    return this._api.getPoints()
-      .then((points) => {
-        points.map((it) => this._store.setItem({key: it.id, item: it.toRAW()}));
-        return points;
-      });
+    if (this._isOnline()) {
+      return this._api.getPoints()
+        .then((points) => {
+          points.map((it) => this._store.setItem({key: it.id, item: it.toRAW()}));
+          return points;
+        });
+    } else {
+      const rawPointsMap = this._store.getAll();
+      const rawPoints = objectToArray(rawPointsMap);
+      const points = ModelPoint.parsePoints(rawPoints);
+      return Promise.resolve(points);
+    }
   }
 
   getDestinations() {
